refactor: change unit test
